requirements.txt
Django>=5.0.4
djangorestframework>=3.16.0 